added changes requested in latest PR and fixed iden3Integration tests
