--- android/coreapiintegration/src/main/java/com/iden3/coreapiintegration/MainActivity.kt
@@ -1,11 +1,0 @@
-package com.iden3.coreapiintegration
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
