+++ android/iden3/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.iden3.iden3" >
+    <uses-permission android:name="android.permission.INTERNET"/>
+</manifest>
