+++ android/iden3/src/main/java/com/iden3/iden3/Iden3Credential.kt
@@ -1,0 +1,12 @@
+package com.iden3.iden3
+
+class Iden3Credential(private var key: String, private var credential: String) {
+
+    fun getKey() : String {
+        return key
+    }
+
+    fun getCredential() : String {
+        return credential
+    }
+}
