+++ android/iden3/src/main/java/com/iden3/iden3/Iden3EventCallback.kt
@@ -1,0 +1,7 @@
+package com.iden3.iden3
+
+interface Iden3EventCallback {
+
+    fun onEventReceived(event: Iden3Event)
+
+}
