+++ android/iden3/src/main/java/com/iden3/iden3/Iden3ProveClaimCallback.kt
@@ -1,0 +1,11 @@
+package com.iden3.iden3
+
+import java.lang.Exception
+
+interface Iden3ProveClaimCallback {
+
+    fun onVerifierResponse(verified: Boolean)
+
+    fun onError(error: Exception)
+
+}
