+++ android/iden3/src/main/java/com/iden3/iden3/Iden3TicketCallback.kt
@@ -1,0 +1,11 @@
+package com.iden3.iden3
+
+import java.lang.Exception
+
+interface Iden3TicketCallback {
+
+    fun onTicketReceived(ticket: Iden3Ticket)
+
+    fun onError(error: Exception)
+
+}
