--- android/iden3coreapi/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.iden3.iden3coreapi" >
-    <uses-permission android:name="android.permission.INTERNET"/>
-</manifest>
