--- android/iden3coreapi/src/main/java/com/iden3/iden3coreapi/Iden3Event.kt
@@ -1,24 +1,0 @@
-package com.iden3.iden3coreapi
-
-import iden3mobile.*
-import java.lang.Exception
-
-class Iden3Event(private var event: Event) {
-
-    fun getTicketId() : String {
-        return event.ticketId
-    }
-
-    fun getData() : String {
-        return event.data
-    }
-
-    fun getType() : String {
-        return event.type
-    }
-
-    fun getError() : Exception {
-        return event.err
-    }
-
-}
