--- android/iden3coreapi/src/main/java/com/iden3/iden3coreapi/Iden3EventListener.kt
@@ -1,7 +1,0 @@
-package com.iden3.iden3coreapi
-
-interface Iden3EventListener {
-
-    fun onEventReceived(event: Iden3Event)
-
-}
