--- android/iden3coreapi/src/main/java/com/iden3/iden3coreapi/Iden3ProveClaimListener.kt
@@ -1,11 +1,0 @@
-package com.iden3.iden3coreapi
-
-import java.lang.Exception
-
-interface Iden3ProveClaimListener {
-
-    fun onClaimProofReceived(proved: Boolean)
-
-    fun onError(error: Exception)
-
-}
