--- android/iden3coreapi/src/main/java/com/iden3/iden3coreapi/Iden3TicketListener.kt
@@ -1,11 +1,0 @@
-package com.iden3.iden3coreapi
-
-import java.lang.Exception
-
-interface Iden3TicketListener {
-
-    fun onTicketReceived(ticket: Iden3Ticket)
-
-    fun onError(error: Exception)
-
-}
