+++ android/iden3integration/proguard-rules.pro
