+++ android/iden3integration/src/main/java/com/iden3/iden3integration/MainActivity.kt
@@ -1,0 +1,11 @@
+package com.iden3.iden3integration
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+}
