+++ android/settings.gradle
@@ -1,3 +1,3 @@
-include ':iden3coreapi'
+include ':iden3'
 include ':gointegration'
-include ':coreapiintegration'
+include ':iden3integration'
