+++ ios/iden3/iden3/Iden3Credential.swift
@@ -1,0 +1,9 @@
+//
+//  Iden3Credential.swift
+//  iden3
+//
+//  Created by Crypto-X on 16/06/2020.
+//  Copyright © 2020 Iden3. All rights reserved.
+//
+
+import Foundation
